Change #1
In this first commit, I have essentially created three ellipses, and
adjusted the speed, and I have pointed the ellipses and told them to
move based on the width and the height of the screen. As well as that,
I wrote in a function of void Boundaries, which makes the if statements
true. Along with that, in the beginning, I have made the boolean
collision true. Which is essentially stating, that when either of the
circles touch the boundaries, the collision should occur.