Change #2
I changed the speed of the third circle I created, to make it more
visible. As well as that, I made the background white, in order for me
to differentiate between the different circles, because I couldn’t
figure out which one was which. Changing the color of each of them
makes it easier to make changes. I also found a fault in the code, as
the if statement wasn’t really working for the collision, I had to make
each of the circles individual in terms of that.