Change #4
In this edition of the code, I changed a lot in terms of the
sizes/speeds and the boolean function. For the sizes, I wanted to make
it look undetected, so I made the original circles look like lines
essentially. As well as that, I added true and false statements in
order to show flashing circles and to draw two or more circles. I also
changed the background colour to a more neutral colour. The true and
false statements in the beginning also make it so that when the
background statement is true, draw 2 circles and to make the circles
keep flashing until the statement is false. I did that for three of the
ellipses.